This update should make it substantially easier to visualize room pieces that intersect with a given area.
Added: AllObjectIntersector to visualize which game room pieces are supposed to be intersecting with the min max points. Makes it a lot easier than individually picking each gameobject and seeing if it intersects or not. Has multiple modes to determine what kind of visualization you want to see (room pieces, just the min max area, or both (with and without boxes)).
Bug: When switching the visualization style, no lines show up. No idea why....
Updated: BoundingBoxVisualizer to make it easier to reach into the class.
Updated PC Scene to use the TrueVisualizer instead of the old one. The old one is still there if you want to toggle it on, toggle the TrueVisualizer off, and see manual intersections.